routing: move failure interpretation into mission control
